Dương

+++ Arkanoid/src/base/MovableObject.java
@@ -1,10 +1,10 @@
 package base;
 public abstract class MovableObject extends GameObject{
-    protected  int dx, dy;
+    protected  double dx, dy;
 
     public MovableObject(){};
 
-    public MovableObject(int x, int y, int width, int height, int dx, int dy) {
+    public MovableObject(int x, int y, int width, int height, double dx, double dy) {
         super(x,y,width,height);
         this.dx = dx;
         this.dy = dy;
